Makefile: curl w/o progress meter, catches HTTP errors
curl normally puts the response body in the -o optarg even
if it hits a 4xx or 5xx error; -f prevents this behavior.
we don't want a 404 page stored under the base.tar.xz name.

the progress meter messes up logs when this is run w/o
a terminal, like in a CI system.

the $(DLR), $(DLR_FLAGS) indirection was not very useful,
what with the hardcoded -o flags.